Refactor sensual processor and instructions for IHL/Sensual integration
- Replace sensual940Processor with IHLsensual940Processor in processors.js and instructions/index.js.
- Remove deprecated sensual.js files from both processors and instructions directories.
- Update company selection in index.js to reflect the new IHL/Sensual designation.
