+++ js/instructions/940/IHLsensual.js
